Add CRAM-MD5 support for SMTP to email app.
The build in email client does not support CRAM-MD5 authentification for
either SMTP or IMAP, so it is not possible to log into servers with AUTH
PLAIN/LOGIN disabled. This patch adds support for CRAM-MD5 for SMTP (but
not for IMAP).

A request for this feature has been unanswered for three years now:
https://code.google.com/p/android/issues/detail?id=5901

Update 1: Changed new line in JavaDoc

Change-Id: I93908741d0c0123b50ac6f8684fe3c81d9d2c45a